refactor(lanes): split lanes.rs into lanes/ module
Purely mechanical extraction of src/lanes.rs (2,970 lines, 87KB) into
src/lanes/mod.rs plus seven focused submodules with zero logic changes.
Public API (LaneAction, LaneDef, run, run_for_pre_release) remains
importable from crate::lanes:: unchanged.
